AI-Powered-Product-Finder/Product-Finder-Alloydb/Backend/Chatbot-MCP/Chatbot-MCP-Client/requirements.txt
langgraph==1.0.4
langchain_google_genai==4.2.0
fastapi==0.121.2
uvicorn==0.35.0
python-dotenv==1.1.1
langchain-mcp-adapters==0.2.1